Remove `service_set_id is null` condition to objects in ObjectTableService.php
For number of objects in ServiceTemplateUsageTable and objects in ObjectsTable to be consistent
`service_set_id is null` condition is removed in prepareQuery() in ObjectTableService.php and left join
on icinga_service_set is added to get the services from service set. Also, the header for object table is
changed from 'Host' to 'Host | Service Set'.